FEAT: Added API Key with Secret
