+++ .gitignore
@@ -41,4 +41,4 @@ fastlane/metadata
 
 ### Project ###
 fastlane/.temp
-
+TBWeather/Resource/Properties/Secrets.plist
