+++ TBWeather/Source/Network/Interface/Services/ServiceWeatherInterface.swift
@@ -1,0 +1,12 @@
+//
+//  ServiceWeatherInterface.swift
+//  TBWeather
+//
+//  Created by Túlio Bazan da Silva on 12/10/22.
+//
+
+import Foundation
+
+protocol WeatherServiceInterface {
+    func getWeatherForecast(at coordinate: Coordinate, unit: MeasurementUnit?) async throws -> WeatherForecast
+}
